- fixed printf warning (thanks stahta01)
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@7117 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/codecompletion/parser/token.h
@@ -88,12 +88,12 @@ public:
         {
             const long totalTime = it->first->m_StopWatch.Time();
             wxString log;
-            log.Printf(_T("\"%s\" used time is %d minute(s), %ld.%03ld seconds; call times is %d."),
+            log.Printf(_T("\"%s\" used time is %ld minute(s), %ld.%03ld seconds; call times is %lu."),
                        it->second.wx_str(),
                        (totalTime / 60000),
                        (totalTime / 1000) % 60,
                        (totalTime % 1000),
-                       it->first->m_CallTimes);
+                       static_cast<unsigned long>(it->first->m_CallTimes));
 #ifndef CC_PARSER_TEST
             Manager::Get()->GetLogManager()->DebugLog(log);
 #endif
